+++ lib/app/presentation/screens/concert_info_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:rock_n_roll_forecast/app/core/theme/app_colors.dart';
 import 'package:rock_n_roll_forecast/app/core/utilities/constants.dart';
 import 'package:rock_n_roll_forecast/app/presentation/cubits/weather/weather_cubit.dart';
 import 'package:rock_n_roll_forecast/app/presentation/widgets/concert_title.dart';
 
-import '../../core/theme/app_colors.dart';
 import '../../core/theme/text_theme.dart';
 import '../../core/utilities/helpers/responsive_helper.dart';
 import '../../core/utilities/helpers/widget_helper.dart';
@@ -12,8 +12,8 @@ import '../cubits/forecast/forecast_cubit.dart';
 import '../cubits/forecast/forecast_state.dart';
 import '../cubits/weather/weather_state.dart';
 import '../widgets/city_overview_loader.dart';
-import '../widgets/daily_forecast.dart';
 import '../widgets/forecast_loader.dart';
+import '../widgets/weather_forecast.dart';
 import '../widgets/weather_overview.dart';
 
 class ConcertInfoScreen extends StatefulWidget {
@@ -49,15 +49,14 @@ class _ConcertInfoScreenState extends State<ConcertInfoScreen> {
     final isMobile = Responsive.isMobile;
 
     return Scaffold(
-      body: Column(
-        children: [
-          ConcertTitle(city: city!),
-          Padding(
-            padding: EdgeInsets.symmetric(
-              horizontal: isMobile ? 18 : 36,
-            ),
-            child: Center(
-              child: BlocBuilder<WeatherCubit, WeatherState>(
+      body: SafeArea(
+        child: SingleChildScrollView(
+          padding: EdgeInsets.all(isMobile ? 18 : 36),
+          child: Column(
+            children: [
+              ConcertTitle(city: city!),
+              const SizedBox(height: 26),
+              BlocBuilder<WeatherCubit, WeatherState>(
                 builder: (context, state) {
                   if (state is WeatherLoading) {
                     return const OverviewLoader();
@@ -80,40 +79,31 @@ class _ConcertInfoScreenState extends State<ConcertInfoScreen> {
                   }
                 },
               ),
-            ),
+              const SizedBox(height: 32),
+              BlocBuilder<ForecastCubit, ForecastState>(
+                builder: (context, state) {
+                  if (state is ForecastLoading) {
+                    return const ForecastLoader();
+                  } else if (state is ForecastLoaded) {
+                    return WeatherForecast(forecast: state.result);
+                  } else if (state is ForecastError) {
+                    return WidgetHelper.error(state.message);
+                  } else {
+                    return WidgetHelper.error(Constants.unknownError);
+                  }
+                },
+              ),
+              const SizedBox(height: 2),
+              Text(
+                "Weather for $city",
+                style: AppTextTheme.textTheme.bodyMedium?.copyWith(
+                  fontWeight: FontWeight.w500,
+                  color: AppColors.grey,
+                ),
+              ),
+            ],
           ),
-          const SizedBox(height: 30),
-          BlocBuilder<ForecastCubit, ForecastState>(
-            builder: (context, state) {
-              if (state is ForecastLoading) {
-                return const ForecastLoader();
-              } else if (state is ForecastLoaded) {
-                return DailyForecast(forecast: state.result);
-              } else if (state is ForecastError) {
-                return Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 18),
-                  child: WidgetHelper.error(state.message),
-                );
-              } else {
-                return Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 18),
-                  child: WidgetHelper.error(
-                    Constants.unknownError,
-                  ),
-                );
-              }
-            },
-          ),
-          SizedBox(height: isMobile ? 22 : 42),
-          Text(
-            "Forecast for $city",
-            style: AppTextTheme.textTheme.bodyLarge?.copyWith(
-              fontWeight: FontWeight.w500,
-              color: AppColors.darkGray,
-            ),
-          ),
-          const SizedBox(height: 32),
-        ],
+        ),
       ),
     );
   }
